+++ src/reactpy/asgi/__init__.py
@@ -1,5 +1,5 @@
-from reactpy.asgi.executors.pyscript import ReactPyCSR
+from reactpy.asgi.executors.pyodide import ReactPyPyodide
 from reactpy.asgi.executors.standalone import ReactPy
 from reactpy.asgi.middleware import ReactPyMiddleware
 
-__all__ = ["ReactPy", "ReactPyCSR", "ReactPyMiddleware"]
+__all__ = ["ReactPy", "ReactPyMiddleware", "ReactPyPyodide"]